Maybe fixed Bitcoin Halving timer getting stuck

+++ src/main/java/eu/time/discordbot/discord/commands/text/BitcoinHalvingCommand.java
@@ -1,6 +1,6 @@
 package eu.time.discordbot.discord.commands.text;
 
-import eu.time.discordbot.crypto.BitcointHalvin;
+import eu.time.discordbot.crypto.BitcointHalving;
 import eu.time.discordbot.discord.command.Permission;
 import eu.time.discordbot.discord.command.TextCommand;
 import net.dv8tion.jda.api.entities.MessageEmbed;
@@ -15,7 +15,13 @@ public class BitcoinHalvingCommand extends TextCommand {
 
     @Override
     protected void execIntern(MessageReceivedEvent event, List<String> args) {
-        MessageEmbed untilHalvingMessage = new BitcointHalvin().getUntilHalvingMessage();
+        MessageEmbed untilHalvingMessage = new BitcointHalving().getUntilHalvingMessage();
+
+        if (untilHalvingMessage == null) {
+            event.getMessage().reply("ERROR: Could not get last Block").queue();
+            return;
+        }
+
         messageHandler.sendEmbed(event, untilHalvingMessage);
         event.getMessage().delete().queue();
     }
